backend/django_login/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
boto3
django-cors-headers
jwt
